Gets rids of a circular dependency between authentication and util. - CoreDefaultsManager is moved to the authentication module. - A backwards compatible reference is inserted in the original util module.
