+++ python/tank/authentication/__init__.py
@@ -21,4 +21,5 @@ credentials are reused if available.
 from .errors import ShotgunAuthenticationError, AuthenticationError, IncompleteCredentials, AuthenticationCancelled
 from .shotgun_authenticator import ShotgunAuthenticator
 from .defaults_manager import DefaultsManager
+from .core_defaults_manager import CoreDefaultsManager
 from .user import ShotgunUser, deserialize_user, serialize_user
