+++ python/tank/util/__init__.py
@@ -17,8 +17,6 @@ from .shotgun import create_event_log_entry
 from .shotgun import get_entity_type_display_name
 from .shotgun import get_published_file_entity_type
 
-from .defaults_manager import CoreDefaultsManager
-
 from .environment import append_path_to_env_var
 from .environment import prepend_path_to_env_var
