+++ python/tank/util/shotgun/connection.py
@@ -16,6 +16,7 @@ from __future__ import with_statement
 
 import os
 import threading
+import urlparse
 
 # use api json to cover py 2.5
 from tank_vendor import shotgun_api3
@@ -200,11 +201,41 @@ def __parse_config_data(file_data, user, shotgun_cfg_path):
         # Make sure it is None.
         config_data["app_store_http_proxy"] = None
 
+    config_data["host"] = cleanup_host(config_data["host"])
+
     return config_data
 
 
+def cleanup_host(server_url):
 
-    
+    # First clean up any extra spaces.
+    server_url = server_url.strip()
+
+    # Then break up the url into chunks
+    parsed_url = urlparse.urlparse(server_url)
+
+    # The given url https://192.168.1.250:30/path?a=b is parsed such that
+    # scheme => https
+    # netloc => 192.168.1.250:30
+    # path = /path
+    # query = a=b
+
+    # As such, when sanitizing a url, we want to keep only the scheme and
+    # network location
+
+    # Then extract the good parts from the url
+    clean_url = urlparse.ParseResult(
+        # We want https when there is no specified scheme.
+        scheme=parsed_url.scheme or "https",
+        # If only a host has been provided, path will be set.
+        # If a scheme was set, then use the netloc
+        netloc=parsed_url.netloc or parsed_url.path,
+        path="", params="", query="", fragment=""
+    )
+
+    return urlparse.urlunparse(clean_url)
+
+
 def get_associated_sg_base_url():
     """
     Returns the shotgun url which is associated with this Toolkit setup.
