+++ tests/authentication_tests/test_auth_settings.py
@@ -18,7 +18,7 @@ from mock import patch, Mock, PropertyMock
 from tank_test.tank_test_base import TankTestBase
 from tank_test.tank_test_base import setUpModule # noqa
 
-from tank.util import CoreDefaultsManager
+from tank.authentication import CoreDefaultsManager
 from tank.authentication import DefaultsManager
 from tank.util.user_settings import UserSettings
